requirements.txt
alembic==1.0.10
APScheduler==3.6.0
astroid==2.2.5
bcrypt==3.1.7
blinker==1.4
cffi==1.12.3
Click==7.0
coverage==4.5.3
dnspython==1.16.0
email-validator==1.0.4
Flask==1.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.19.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Session
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
idna==2.8
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
Mako==1.0.12
MarkupSafe==1.1.1
marshmallow==2.19.5
mccabe==0.6.1
pep8==1.7.1
psycopg2==2.8.3
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-usernames==0.2.2
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.5
typed-ast==1.4.0
tzlocal==1.5.1
webargs==5.3.2
Werkzeug==0.15.4
wrapt==1.11.2
